Fixed creation of communication service
